[Add] NPC가 피자 사이즈, 상태에 따른 돈 지불과 벌금 기능 추가

+++ Assets/Scripts/Manger/ShopManger.cs
@@ -39,7 +39,7 @@ public class ShopManger : Singleton<ShopManger>
     }
     public bool CheckPlayerMoney(string itemName)
     {
-        int playerMoney = PlayerController.Instance.PlayerMoney;
+        float playerMoney = PlayerController.Instance.PlayerMoney;
 
         if (playerMoney <= 0)
         {
